+++ src/main/java/org/betonquest/betonquest/api/bukkit/events/LoadDataEvent.java
@@ -1,4 +1,4 @@
-package org.betonquest.betonquest.api;
+package org.betonquest.betonquest.api.bukkit.events;
 
 import org.bukkit.event.Event;
 import org.bukkit.event.HandlerList;
