+++ src/main/java/org/betonquest/betonquest/api/bukkit/events/QuestCompassTargetChangeEvent.java
@@ -1,4 +1,4 @@
-package org.betonquest.betonquest.api;
+package org.betonquest.betonquest.api.bukkit.events;
 
 import org.betonquest.betonquest.api.profiles.Profile;
 import org.betonquest.betonquest.api.profiles.ProfileEvent;
