+++ src/main/java/org/betonquest/betonquest/conversation/Conversation.java
@@ -4,9 +4,9 @@ import net.md_5.bungee.api.chat.BaseComponent;
 import net.md_5.bungee.api.chat.TextComponent;
 import org.apache.commons.lang3.tuple.Pair;
 import org.betonquest.betonquest.BetonQuest;
-import org.betonquest.betonquest.api.ConversationOptionEvent;
-import org.betonquest.betonquest.api.PlayerConversationEndEvent;
-import org.betonquest.betonquest.api.PlayerConversationStartEvent;
+import org.betonquest.betonquest.api.bukkit.events.ConversationOptionEvent;
+import org.betonquest.betonquest.api.bukkit.events.PlayerConversationEndEvent;
+import org.betonquest.betonquest.api.bukkit.events.PlayerConversationStartEvent;
 import org.betonquest.betonquest.api.config.quest.QuestPackage;
 import org.betonquest.betonquest.api.logger.BetonQuestLogger;
 import org.betonquest.betonquest.api.profiles.OnlineProfile;
