+++ src/main/java/org/betonquest/betonquest/database/PlayerData.java
@@ -6,9 +6,9 @@ import org.betonquest.betonquest.Journal;
 import org.betonquest.betonquest.Point;
 import org.betonquest.betonquest.Pointer;
 import org.betonquest.betonquest.api.Objective;
-import org.betonquest.betonquest.api.PlayerTagAddEvent;
-import org.betonquest.betonquest.api.PlayerTagRemoveEvent;
-import org.betonquest.betonquest.api.PlayerUpdatePointEvent;
+import org.betonquest.betonquest.api.bukkit.events.PlayerTagAddEvent;
+import org.betonquest.betonquest.api.bukkit.events.PlayerTagRemoveEvent;
+import org.betonquest.betonquest.api.bukkit.events.PlayerUpdatePointEvent;
 import org.betonquest.betonquest.api.logger.BetonQuestLogger;
 import org.betonquest.betonquest.api.profiles.Profile;
 import org.betonquest.betonquest.config.Config;
