+++ src/main/java/org/betonquest/betonquest/quest/event/compass/CompassEvent.java
@@ -1,7 +1,7 @@
 package org.betonquest.betonquest.quest.event.compass;
 
 import org.betonquest.betonquest.BetonQuest;
-import org.betonquest.betonquest.api.QuestCompassTargetChangeEvent;
+import org.betonquest.betonquest.api.bukkit.events.QuestCompassTargetChangeEvent;
 import org.betonquest.betonquest.api.config.quest.QuestPackage;
 import org.betonquest.betonquest.api.logger.BetonQuestLogger;
 import org.betonquest.betonquest.api.profiles.Profile;
